+++ dbupgrade/__init__.py
@@ -1,0 +1,6 @@
+from .upgrade import db_upgrade as db_upgrade  # noqa: F401
+from .version import (  # noqa: F401
+    MAX_API_LEVEL as MAX_API_LEVEL,
+    MAX_VERSION as MAX_VERSION,
+    VersionInfo as VersionInfo,
+)